% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsprec.R
\name{jags.prec}
\alias{jags.prec}
\title{Returns and/or plots relative precision of different confidence levels from jags output.}
\usage{
jags.prec(jags.out, plot = T, conf = c(0.5, 0.9, 0.95), legend = T,
  return.output = F, which.param = NULL)
}
\arguments{
\item{jags.out}{A jags output object from R2jags}

\item{plot}{Whether or not to produce a plot of precision levels.  Defaults to TRUE.}

\item{conf}{A vector of confidence levels to report.  Defaults to c(.5,.9,.95)}

\item{legend}{Whether or not to produce a legend, if plot==T.}

\item{return.output}{Whether or not to produce an output table.  Defaults to FALSE.}

\item{which.param}{A vector of the parameter indices we're interested in plotting.  Calling jags.names() first can help with this.}
}
\description{
Returns and/or plots relative precision of different confidence levels from jags output.
}
\examples{
# ------------ a sample R2jags model ----------- #

\dontrun{
library(R2jags)
N <- 1000
x <- 1:N
epsilon <- rnorm(N, 0, 1)
y <- 5 + x + epsilon

cat('model {
 for (i in 1:N){
   y[i] ~ dnorm(y.hat[i], tau)
   y.hat[i] <- a + b * x[i]
 }
 a ~ dnorm(0, .0001)
b ~ dnorm(0, .0001)
tau <- pow(sigma, -2)
sigma ~ dunif(0, 100)
}', file="test.jags")

test.data <- list(x=x,y=y,N=N)
test.jags.out <- jags(model.file="test.jags", data=test.data, 
     parameters.to.save=c("a","b","tau","sigma"), n.chains=3, n.iter=1000)

# ------------- calling jags.prec() ------------ #
jags.names(test.jags.out)
jags.prec(test.jags.out,plot=T,return.output=T)
jags.prec(test.jags.out,plot=T,return.output=F,conf=c(.5,.9),which.param=c(1:2,4:5))
}
}
\author{
Matt Tyers
}

